requirements.txt
requests
mysql-connector-python
schedule